html-cdn-py/pyproject.toml
[project]
name = "quotient-examples-html-cdn-py"
version = "0.1.0"
description = "quotient-examples-html-cdn-py -- generic example using quotientjs from unpkg"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv",
    "pydantic",
    
    "uvicorn",
    "fastapi",
    
    "jinja2",
    
    "setuptools",
    "watchfiles",
]